feat: move compose to feed
